+++ gba_test/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+### Changed
+- `ShouldPanic::YesWithMessage(message)` no longer requires dynamic allocation.
 ### Fixed
 - `ShouldPanic::YesWithMessage(message)` now matches against the panic message itself, not including location info.